+++ internal/infrastructure/sql/schema.sql
@@ -1,15 +1,15 @@
 CREATE TABLE channels (
   id SERIAL PRIMARY KEY,
   code TEXT NOT NULL,
-  channel_chat_id INTEGER,
-  admin_chat_id INTEGER,
-  discussions_chat_id INTEGER,
+  channel_chat_id BIGINT,
+  admin_chat_id BIGINT,
+  discussions_chat_id BIGINT,
   decorations TEXT
 );
 
 CREATE TABLE bots (
   id SERIAL PRIMARY KEY,
-  tgid INTEGER NOT NULL,
+  tgid BIGINT NOT NULL,
   type TEXT NOT NULL,
   channel_id INTEGER REFERENCES channels(id),
